stori/pyproject.toml
[tool.poetry]
name = "stori"
version = "0.1.0"
description = "A technical test for a Python Backend job at Stori. Its the backend."
authors = ["Armando Ezequiel Puerta <armando.ezequiel.puerta@gmail.com>"]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/EzequielPuerta-TechnicalTests/stori"
repository = "https://github.com/EzequielPuerta-TechnicalTests/stori"
documentation = "https://github.com/EzequielPuerta-TechnicalTests/stori"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
django-ckeditor = "^6.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
pre-commit = "^3.4.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
isort = "^5.12.0"
black = {version = "^23.9.1", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.git
    | \.mypy_cache
    | build
    | dist
  )/
  | stori/settings.py
)
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py"]
